# api-definition.yml
openapi: 3.0.0
info:
  description: |
    This is the Stereodose API
    [https://github.com/briansimoni/stereodose](https://github.com/briansimoni/stereodose)
  version: '1.0.0'
  title: Stereodose API
  license:
    name: GPL 3.0
    url: 'https://github.com/briansimoni/stereodose/blob/master/LICENSE'
servers:
  - url: 'https://stereodose.app'
  - url: 'http://localhost:3000'
  - url: 'http://localhost:4000'
paths:
...
  '/api/playlists/':
    get:
      tags:
        - playlists
      summary: get a set of playlists
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
        - in: query
          name: limit
          schema:
            type: integer
          description: Limit the of items to return
        - in: query
          name: category
          schema:
            type: string
          description: The overall category of playlist
        - in: query
          name: subcategory
          schema:
            type: string
          description: The subcategory of playlist
        - in: query
          name: sort-key
          schema:
            type: string
          description: "Can be one of the following defaulting to likes_count: created_at, updated_at, name, likes_count"
        - in: query
          name: order
          schema:
            type: string
          description: "Can be one of the following defaulting to desc: asc, desc"
        - in: query
          name: spotify-ids
          schema:
            type: string
          description: "A space delimited set of playlist IDs"
      responses:
        200:
          description: A JSON array of playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'

    post:
      tags:
        - playlists
      summary: create a playlist
      security:
        - cookieAuth: []
      requestBody:
        description: Create a new Stereodose playlist
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                SpotifyID:
                  type: string
                Category:
                  type: string
                Subcategory:
                  type: string
                ImageURL:
                  type: string
                ThumbnailURL:
                  type: string
      responses:
        201:
          description: successfully created playlist
...
  '/api/playlists/{id}':
    get:
      tags:
        - playlists
      summary: get a playlist
      parameters:
        - in: path
          name: id # Note the name is the same as in the path
          required: true
          schema:
            type: string
          description: The playlist ID
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'

    delete:
      tags:
        - playlists
      summary: delete a playlist
      description: Deletes the playlist and associated likes and comments. Only admins may perform this operation
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The playlist ID
      responses:
        204:
          description: no content indicates success
...
components:
  schemas:
...
    Playlist:
      type: object
      properties:
        spotifyID:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        category:
          type: string
        categoryDisplayName:
          type: string
        subCategory:
          type: string
        collaborative:
          type: boolean
        href:
          type: string
        name:
          type: string
        public:
          type: boolean
        snapshot_id:
          type: string
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        likes:
          type: array
          items:
            $ref: '#/components/schemas/Like'
        likesCount:
          type: integer
        URI:
          type: string
        userId:
          type: integer
        bucketImageURL:
          type: string
        bucketThumbnailURL:
          type: string
        permalink:
          type: string
    Track:
      type: object
      properties:
        spotifyID:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        name:
          type: string
        duration:
          type: integer
        previewURL:
          type: string
        trackNumber:
          type: string
        URI:
          type: string
        artists:
          type: string
        spotifyArtistIDs:
          type: string
    Comment:
      type: object
      properties:
        ID:
          type: integer
        CreatedAt:
          type: string
        UpdatedAt:
          type: string
        DeletedAt:
          type: string
        content:
          type: string
        PlaylistID:
          type: string
        userID:
          type: string
        displayName:
          type: string
        permalink:
          type: string
    Like:
      type: object
      properties:
        ID:
          type: string
        CreatedAt:
          type: string
        UpdatedAt:
          type: string
        userID:
          type: integer
        playlistID:
          type: string
        playlistName:
          type: string
        playlist: 
          $ref: '#/components/schemas/Playlist'
        permalink:
          type: string
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: stereodose_session